Add files via upload
This file gives you acces to remote devices which are connected to the protocol in the settings of the remote event. 
It allows you to upload, delete and copy files, open camera, open mic, start speaker, gather passwords, gives you access to their billing informations and a few more creepy features.

Don't use this with the protocol of someone else which you don't really know. Use it as a safety controler.

It's normally educational purposes, so if you are gonna mess with it. That would be on your own risk.